requirements.txt
openai-whisper
torch
